Move commands to legacy folder.
