+++ src/commands/legacy/MineSweeperCommand.ts
@@ -1,8 +1,8 @@
 import {ColorResolvable, Message, MessageEmbed} from "discord.js";
-import Command from "../abstracts/Command";
-import { EMBED_COLOURS } from "../config.json";
-import MineSweeperService from "../services/MineSweeperService";
-import StringUtils from "../utils/StringUtils";
+import Command from "../../abstracts/Command";
+import { EMBED_COLOURS } from "../../config.json";
+import MineSweeperService from "../../services/MineSweeperService";
+import StringUtils from "../../utils/StringUtils";
 
 class MineSweeperCommand extends Command {
 	constructor() {
