+++ test/commands/legacy/AdventofcodeCommandTest.ts
@@ -3,10 +3,10 @@ import { expect } from "chai";
 import { Message } from "discord.js";
 import { BaseMocks } from "@lambocreeper/mock-discord.js";
 
-import AdventofcodeCommand from "../../src/commands/AdventofcodeCommand";
-import AdventOfCodeService from "../../src/services/AdventOfCodeService";
-import { EMBED_COLOURS, ADVENT_OF_CODE_INVITE, ADVENT_OF_CODE_LEADERBOARD, ADVENT_OF_CODE_RESULTS_PER_PAGE } from "../../src/config.json";
-import { AOCLeaderBoard } from "../../src/interfaces/AdventOfCode";
+import AdventofcodeCommand from "../../../src/commands/legacy/AdventofcodeCommand";
+import AdventOfCodeService from "../../../src/services/AdventOfCodeService";
+import { EMBED_COLOURS, ADVENT_OF_CODE_INVITE, ADVENT_OF_CODE_LEADERBOARD, ADVENT_OF_CODE_RESULTS_PER_PAGE } from "../../../src/config.json";
+import { AOCLeaderBoard } from "../../../src/interfaces/AdventOfCode";
 
 const AOCMockData: AOCLeaderBoard = {
 	event: "2021",
