+++ test/commands/legacy/ProjectCommandTest.ts
@@ -2,9 +2,9 @@ import { expect } from "chai";
 import { Message } from "discord.js";
 import { createSandbox, SinonSandbox } from "sinon";
 import { BaseMocks } from "@lambocreeper/mock-discord.js";
-import Command from "../../src/abstracts/Command";
-import { EMBED_COLOURS } from "../../src/config.json";
-import ProjectCommand from "../../src/commands/ProjectCommand";
+import Command from "../../../src/abstracts/Command";
+import { EMBED_COLOURS } from "../../../src/config.json";
+import ProjectCommand from "../../../src/commands/legacy/ProjectCommand";
 
 describe("ProjectCommand", () => {
 	describe("constructor", () => {
