+++ test/commands/legacy/ResourcesCommandTest.ts
@@ -3,8 +3,8 @@ import { SinonSandbox, createSandbox } from "sinon";
 import { Message } from "discord.js";
 import { BaseMocks } from "@lambocreeper/mock-discord.js";
 
-import Command from "../../src/abstracts/Command";
-import ResourcesCommand from "../../src/commands/ResourcesCommand";
+import Command from "../../../src/abstracts/Command";
+import ResourcesCommand from "../../../src/commands/legacy/ResourcesCommand";
 
 describe("ResourcesCommand", () => {
 	describe("constructor()", () => {
